Added LL Debugger config + win detection v0.1

+++ src/controllers/console_controller.rs
@@ -1,12 +1,12 @@
 use crate::views::console_view;
 use std::io::stdin;
 
-pub fn read_col() -> u8 {
+pub fn read_col() -> usize {
     loop {
         let mut buf = String::new();
         stdin().read_line(&mut buf).unwrap();
         buf = buf.replace("\r\n", "");
-        let input: u8 = match buf.parse() {
+        let input: usize = match buf.parse() {
             Result::Ok(r) => r,
             Result::Err(_) => {
                 console_view::print_msg("Please enter a column number between 1 and 7: ");
